Update patch set 1
Patch Set 1:

> I'll fix this for now, but I'd like to know what RDO's plans for
 > python2 support are :-)

Thanks! We're waiting for the CentOS community to have all the required bits ready to start building CentOS8-based packages. Once that is ready, we will drop CentOS 7 support for the Ussuri release. So it should be a matter of a couple months, if everything goes as expected.

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0